Gpt Python Jupyter Notebook
Some experiments with GPT getting it to draw me charts and figures.
Asked it to make me an animation, and it made a kind of sine wave.
These are just the "drawings" it made.
Also I got it to write me a program to rotate a moon and orbit around a planet in Blender, I was trying to get it to design some vertex shaders but it was getting supercomplicated. I tried Julia and ran into issues downloading packages, so as a last resort I tried Python. I was also able to get it to "draw" using stable diffusion, by getting it to write prompts.

For some reason, most people seem to be using it by checking what it knows already, rather than teaching it the information it needs in order to make the decision. In this way I was able to get it to write using a dictionary of Nuspeak.

Epistemological Experiments
"How do we know what we know?"